==> 03_18.R <==
library(mgcv)
library(xtable)
#lag&move average
COm1<-mvdata(fit_data$CO)
SO2m1<-mvdata(fit_data$SO2)
O3m1<-mvdata(fit_data$O3)
PM2.5m1<-mvdata(fit_data$PM2.5)
PM10m1<-mvdata(fit_data$PM10)
NO2m1<-mvdata(fit_data$NO2)
NOm1<-mvdata(fit_data$NO)
#RAINm<-mvdata(data.air$RAIN)
#TEMPm<-mvdata(data.air$TEMP)
#RHm<-mvdata(data.air$RH)

CO_allergy1<-gamair1(14,COm1,fit_data,TEMPm,RHm)
SO2_allergy1<-gamair1(14,SO2m1,fit_data,TEMPm,RHm)
O3_allergy1<-gamair1(14,O3m1,fit_data,TEMPm,RHm)
PM2.5_allergy1<-gamair1(14,PM2.5m1,fit_data,TEMPm,RHm)
PM10_allergy1<-gamair1(14,PM10m1,fit_data,TEMPm,RHm)
NO_allergy1<-gamair1(14,NOm1,fit_data,TEMPm,RHm)
NO2_allergy1<-gamair1(14,NO2m1,fit_data,TEMPm,RHm)

xtable(CO_allergy1[[1]],digits =3)
xtable(SO2_allergy1[[1]],digits =3)
xtable(O3_allergy1[[1]],digits =3)
xtable(PM2.5_allergy1[[1]],digits =3)
xtable(PM10_allergy1[[1]],digits =3)
xtable(NO_allergy1[[1]],digits =3)
xtable(NO2_allergy1[[1]],digits =3)

xtable(CO_allergy1[[2]],digits =3)
xtable(SO2_allergy1[[2]],digits =3)
xtable(O3_allergy1[[2]],digits =3)
xtable(PM2.5_allergy1[[2]],digits =3)
xtable(PM10_allergy1[[2]],digits =3)
xtable(NO_allergy1[[2]],digits =3)
xtable(NO2_allergy1[[2]],digits =3)

CO_hives1<-gamair1(15,COm1,fit_data,TEMPm,RHm)
SO2_hives1<-gamair1(15,SO2m1,fit_data,TEMPm,RHm)
O3_hives1<-gamair1(15,O3m1,fit_data,TEMPm,RHm)
PM2.5_hives1<-gamair1(15,PM2.5m1,fit_data,TEMPm,RHm)
PM10_hives1<-gamair1(15,PM10m1,fit_data,TEMPm,RHm)
NO_hives1<-gamair1(15,NOm1,fit_data,TEMPm,RHm)
NO2_hives1<-gamair1(15,NO2m1,fit_data,TEMPm,RHm)



xtable(CO_hives1[[1]],digits =3)
xtable(SO2_hives1[[1]],digits =3)
xtable(O3_hives1[[1]],digits =3)
xtable(PM2.5_hives1[[1]],digits =3)
xtable(PM10_hives1[[1]],digits =3)
xtable(NO_hives1[[1]],digits =3)
xtable(NO2_hives1[[1]],digits =3)

xtable(CO_hives1[[2]],digits =3)
xtable(SO2_hives1[[2]],digits =3)
xtable(O3_hives1[[2]],digits =3)
xtable(PM2.5_hives1[[2]],digits =3)
xtable(PM10_hives1[[2]],digits =3)
xtable(NO_hives1[[2]],digits =3)
xtable(NO2_hives1[[2]],digits =3)
